refactor(userspace): extract chisel from field_info
